Fix chunk behavior when new blocks are inserted into middle of palette
This is a fix for several issues which were observed after block placement
was implemented. Most notably, when a block was placed into a chunk which
did not yet have the block in its palette, the block would be inserted
into the middle of the palette (to allow for binary search); however,
the data array would not be updated to account for the fact that
palette entries after the new entry would be offseted by 1.