Add `ChunkEntites::entities_within_view_distance()`
This function efficiently finds all entities within the view
distance of a player. It can be used to, for example, send initial
entities to the player when they join. However, it should not
be used for broadcasting packets; `Util::broadcast` is more
efficient in this regard, since it does not have to iterate
over `view_distance^2` chunks.

See also #78.