Only send entity updates to players able to see them - resolves #105
* Add Util::broadcast() function

This function adds an easy way to broadcast a packet
to all players who are able to see a given chunk.
For example, movement packets for a player can
be sent to all players able to see that player's
chunk.

The function hasn't yet been used, but player action
broadcast systems will soon be modified to use
`Util::broadcast()`.

Also see #78.

* Add `ChunkEntites::entities_within_view_distance()`

This function efficiently finds all entities within the view
distance of a player. It can be used to, for example, send initial
entities to the player when they join. However, it should not
be used for broadcasting packets; `Util::broadcast` is more
efficient in this regard, since it does not have to iterate
over `view_distance^2` chunks.

See also #78.

* Implement view distance-based entity creation and destruction on client