Send chunks closest to a player first
This fixes an issue which was discussed on Discord today.

> This one is rather obscure.
When the client joins, it appears to wait for some time before sending the spawn packets. If the server doesn't send all chunks by the time this happens, the client will spawn itself and assume that the closest chunks have already been sent. However, Feather does not send chunks in any particular order; it just sends them as they are loaded.
The addition of lighting to the chunk load process caused it to take a longer time, which led to not all chunks being sent in time. As a result, the client joins itself, but the chunk it is in has not been sent, so it attempts to teleport itself.
Note these lines in Glowstone https://github.com/GlowstoneMC/Glowstone/blob/2fee1587de0979fea84fc2d67eaf78dd6d504142/src/main/java/net/glowstone/entity/GlowPlayer.java#L1089-L1098