Create world level.dat when it doesn't exist (#141)
Creates a world if the `world` directory doesn't exist.

For now it will only write the level.dat file when creating it for the first time.

I added a config entry to change the world directory feather looks for (defaults to world). There's also options for seed and generator, only used when a new world is being created.