Basic block-to-item mappings (#142)
Note that this does not account for blocks such as stone, where breaking
it with a non-silk touch tool drops cobblestone rather than stone. It
also does not account for the tools needed to break any given block,
so Block::IronOre::to_item() will return Item::IronOre regardless
of the tool used to break it.

See also #139.