Rewrite networking code to use Tokio (#42)
# Summary
* Replaced the mio-based IO event loop with a Tokio task
* Rewrote packet encoding/decoding code, should now be more efficient
* Converted `Packet` to write to `BytesMut` instead of `ByteBuf`; `ByteBuf` removed

# Motivation
The current network code is a homebrew asynchronous IO event loop based on mio. Numerous issues have appeared with this implementation, and the code is quite inefficient as well (three to five HashMap lookups when data is received or sent).

Tokio provides its own heavily optimized event loop, in addition to a much easier to use interface. Furthermore, it can extend the async IO capability beyond the TCP streams themselves; for example, the authentication request to the Mojang API in the initial handler can now be run asynchronously, which improves performance (somewhat).

Ultimately, the decision to switch to Tokio is based on performance and stability, which both are improved through the switch.

# Implications
* Server packet handling code will be unaffected.
* The new networking code uses async/await, which was stabilized a few weeks ago and will be released in Rust stable in early to mid November. In the meantime, compiling Feather will require Rust nightly.

Fixes #34 .
Resolves #35.
Resolves #36.