Change how commands handle successes, implement clear command (#279)
* add gitignore for vscode

* Overhaul command success, also implement /clear

* Remove debug prints, oops

* Fix whitespace

* Remove extra unwraps

> Allows commands to return a Option to provide a success statement, and implements the /clear command.

The ItemArgument class does not support item metadata as none of the rest of feather seems to either, that needs to be implemented. Apart from that this should work identically as to how the clear command works in vanilla 1.13.2 (based on switching between a singleplayer world and feather and comparing).