blocks: Initial port to 1.16
This ports the block code generator to 1.16 and the new crate layout.
I've changed it to ingest data from the Vanilla data report outputted
according to https://wiki.vg/Data_Generators.

Instead of taking data from feather-data, the generator now expects the
blocks report to have been manually generated in the current directory.
The benefit of this is that feather-data can be removed from the build process. I know many had
problems with the JDK and rustfmt requirements, so this simplifies things nicely.

Block code is no longer automatically generated at build time; now creating it
must be done manually. That said, it's cached in Git, so regenerating code
is only necessary when we upgrade to a new version.

This commit only ports the generator. What still needs to be done:
* Block categories, wall_blocks, directions
* SimplifiedBlockKind