Implement plugin messaging (#374)
Adds support for sending plugin messages from plugins.

The `send_plugin_message` function in Quill needs to be moved to a function on a player struct or similar to enforce that the entity is a player. However, as we don't yet have a way to do this (it's on the roadmap), the function is currently exposed on `Game`.