Implement world saving along with a new custom world format.
This commit introduces a custom world format that will be used by default in Feather. It has
a few modernizations over the vanilla Anvil format; see comments for details.

In the future, we can add back support for vanilla worlds for compatibility, of course. libcraft-anvil
just needs to be connected with a WorldSource implementation.